feat: refactor component imports and update navigation links for consistency
