+++ src/app/(dashboard)/systems/components/Columns.tsx
@@ -8,11 +8,11 @@ import {
     DropdownMenu,
     DropdownMenuContent,
     DropdownMenuSeparator,
-    DropdownMenuTrigger
+    DropdownMenuTrigger,
 } from '@/components/ui';
 import type { System } from '@/type/System';
 import type { ColumnDef } from '@tanstack/table-core';
-import {CircleCheckBig, CircleOff, MoreVertical} from 'lucide-react';
+import { CircleCheckBig, CircleOff, MoreVertical } from 'lucide-react';
 
 export const columns: ColumnDef<System>[] = [
     {
@@ -28,9 +28,17 @@ export const columns: ColumnDef<System>[] = [
         header: 'Status',
         cell: ({ row }) => {
             const system = row.original;
-            return system.status
-                ? <Badge variant="outline"> <CircleCheckBig className="text-green-500 dark:text-green-400" /> Active</Badge>
-                : <Badge variant="destructive"> <CircleOff /> Inactive</Badge>;
+            return system.status ? (
+                <Badge variant="outline">
+                    {' '}
+                    <CircleCheckBig className="text-green-500 dark:text-green-400" /> Active
+                </Badge>
+            ) : (
+                <Badge variant="destructive">
+                    {' '}
+                    <CircleOff /> Inactive
+                </Badge>
+            );
         },
     },
     {
