+++ src/app/(dashboard)/systems/page.tsx
@@ -1,7 +1,7 @@
 import HeaderSystems from '@/app/(dashboard)/systems/components/HeaderSystems';
 import { SystemProvider } from '@/app/(dashboard)/systems/components/SystemProvider';
 
-export default async function SystemsPage() {
+export default function SystemsPage() {
     return (
         <>
             <HeaderSystems />
